refactor(core): remove unused widgets
