--- packages/widgetbook_core/lib/src/core/core.dart
@@ -1,1 +1,0 @@
-export 'models/models.dart';
