--- packages/widgetbook_core/lib/src/core/models/model.dart
@@ -1,7 +1,0 @@
-class Model {
-  const Model({
-    required this.id,
-  });
-
-  final String id;
-}
