--- packages/widgetbook_core/lib/src/core/models/models.dart
@@ -1,1 +1,0 @@
-export 'model.dart';
