+++ packages/widgetbook_core/lib/src/navigation_tree/widgets/widgets.dart
@@ -2,3 +2,4 @@ export 'navigation_panel.dart';
 export 'navigation_tree.dart';
 export 'navigation_tree_item.dart';
 export 'navigation_tree_node.dart';
+export 'search_field.dart';
