--- packages/widgetbook_core/lib/src/search/search.dart
@@ -1,1 +1,0 @@
-export 'widgets/search_field.dart';
