--- packages/widgetbook_core/lib/src/settings/models/models.dart
@@ -1,2 +1,0 @@
-export 'setting_section_data.dart';
-export 'settings_panel_data.dart';
