--- packages/widgetbook_core/lib/src/settings/models/setting_section_data.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-class SettingSectionData {
-  const SettingSectionData({
-    required this.name,
-    required this.settings,
-  });
-
-  final String name;
-  final List<Widget> settings;
-}
