--- packages/widgetbook_core/lib/src/settings/models/settings_panel_data.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-class SettingsPanelData {
-  const SettingsPanelData({
-    required this.name,
-    required this.settings,
-  });
-
-  final String name;
-  final List<Widget> settings;
-}
