+++ packages/widgetbook_core/lib/src/settings/setting.dart
