+++ packages/widgetbook_core/lib/src/settings/settings.dart
@@ -1,2 +1,5 @@
-export 'models/models.dart';
-export 'widgets/widgets.dart';
+export 'dropdown_setting.dart';
+export 'knob_property.dart';
+export 'setting.dart';
+export 'setting_header.dart';
+export 'settings_panel.dart';
