--- packages/widgetbook_core/lib/src/settings/widgets/widgets.dart
@@ -1,9 +1,0 @@
-export 'complex_setting.dart';
-export 'dropdown_setting.dart';
-export 'knob_property.dart';
-export 'setting.dart';
-export 'setting_group.dart';
-export 'setting_header.dart';
-export 'setting_section.dart';
-export 'settings_panel.dart';
-export 'sub_setting.dart';
