+++ packages/widgetbook_core/lib/widgetbook_core.dart
@@ -1,6 +1,4 @@
-export 'src/core/core.dart';
 export 'src/icons/icons.dart';
 export 'src/navigation_tree/navigation_tree.dart';
-export 'src/search/search.dart';
 export 'src/settings/settings.dart';
 export 'src/themes/themes.dart';
