--- packages/widgetbook_core/test/src/core/models/model_test.dart
@@ -1,19 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:widgetbook_core/widgetbook_core.dart';
-
-void main() {
-  group(
-    '$Model',
-    () {
-      test(
-        'returns normally',
-        () {
-          expect(
-            () => const Model(id: '1'),
-            returnsNormally,
-          );
-        },
-      );
-    },
-  );
-}
